requirements.txt
tensorflow>=2.14.0
numpy>=1.24.3
pillow
gradio
